fixed health bar issues

+++ PWNTHEON/Assets/Player/Scripts/healthbar.cs
@@ -8,12 +8,12 @@ public class healthbar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void setMaxHealth (int max) {
+    public void setMaxHealth (float max) {
         slider.maxValue = max;
         slider.value = max;
         fill.color = gradient.Evaluate(1f);
     }
-    public void setHealth(int health) {
+    public void setHealth(float health) {
         slider.value = health;
         fill.color = gradient.Evaluate(slider.normalizedValue);
     }
